# assignment 3/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
      "title": "Volcanoes around the world v2",
      "description": "This API has been created to support assignment two in CAB230, Web Computing. It exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below.",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "yongzhi.zhou@connect.qut.edu.au"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0.19"
    },
...
    "servers": [
      {
        "url": "/api/v3"
      }
    ],
...
    "paths": {
      "/countries": {
        "get": {
          "tags": [
            "Data"
          ],
          "description": "Returns a list of all countries that are associated with one or more volcanoes, ordered alphabetically.",
          "responses": {
            "200": {
              "description": "An array of countries. An example of the first five elements in the array is shown below.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/countries"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid query parameters.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                }
                }
              }
            }
          }
          
        }
      },
      "/star/like/{id}": {
        "post": {
          "tags": [
            "Custom endpoints"
          ],
          "description": "Users can like the volcano.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "The ID of the volcano",
              "required": true,
              "schema": {
                "type": "number"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "An array of object star.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/star"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid query parameters.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/profileput401"
                }
                }
              }
            }
          }
          ,
          "security": [
            {
              "BearerAuth": [
                "write:pets",
                "read:pets"
              ]
            }
          ]
          
        }
      },
      "/star/search": {
        "get": {
          "tags": [
            "Custom endpoints"
          ],
          "description": "Return to Volcanoes Liked by Users.",
          "responses": {
            "200": {
              "description": "Return to Volcanoes Liked by Users.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/star"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid query parameters.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                }
                }
              }
            }
          }
          
        }
      },
      "/volcanoes/year/{year}": {
        "get": {
          "tags": [
            "Custom endpoints"
          ],
          "parameters": [
            {
                "name": "year",
                "in": "params",
                "description": "Year of volcano: CE, BCE",
                "schema": {
                  "type": "string",
                  "enum":[
                    "BCE",
                    "CE"
                  ]
                }
              }
          ],
          "description": "Return to the list of all countries associated with one or more volcanoes according to B.C.E as well as C.E..",
          "responses": {
            "200": {
              "description": "An array of volcanoes",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/volcanobyid"
                  }
                }
              }
            },
            "400": {
              "description": "Error in MySQL query.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                }
                }
              }
            }
          }
          
        }
      },
      "/volcanoes": {
        "get": {
          "tags": [
            "Data"
          ],
          "description": "Returns a list of volcanoes that are associated with the queried country. The country query parameter is required. The list can optionally be filtered by using the populatedWithin query parameter. This will return a filtered list of volcanoes that have at least one person living within the provided radius.",
        
          "parameters": [
            {
              "name": "country",
              "in": "query",
              "description": "Name of country",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
                "name": "populatedWithin",
                "in": "query",
                "description": "Distance within X km of a volcano Available values : 5km, 10km, 30km, 100km",
                "schema": {
                  "type": "string",
                  "enum":[
                    "5km",
                    "10km",
                    "30km",
                    "100km"
                  ]
                }
              }
          ],
          "responses": {
            "200": {
              "description": "An array of objects containing id, name, country, region and subregion properties. An example of one object in the array is shown below.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/volcanoes"
                  }
                }
              }
            },
            "400": {
                "description": "Missing country query parameter or query parameters invalid. Click on 'Schema' below to see the possible error responses.",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/ErrorResponsevolcanoes"
                  }
                  }
                }
              }
          }
        }
      },
...
      "/user/register": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "description": "Creates a new user account. A request body containing the user to be registered must be sent.",
          "requestBody": {
            "content": {
              "application/json": {
                
                "schema": {
                    
                    "$ref": "#/components/schemas/register"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "User successfully created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/register201"
                  }
                }
              }
            },
            "400": {
                "description": "Bad request",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/register400"
                  }
                  }
                }
            },
            "409": {
                "description": "User already exists",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/register409"
                  }
                  }
                }
            }
          }
        }
      },
      "/user/login": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "description": "Log in to an existing user account. A request body containing the user credentials must be sent.",
          "requestBody": {
            "content": {
              "application/json": {
                
                "schema": {
                    
                    "$ref": "#/components/schemas/register"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Log in successful",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/login200"
                  }
                }
              }
            },
            "400": {
                "description": "Invalid log in request",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/login400"
                  }
                  }
                }
            },
            "401": {
                "description": "Log in failed",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/login401"
                  }
                  }
                }
            }
          }
        }
      },
      "/user/{email}/profile": {
        "get": {
          "tags": [
            "Profile"
          ],
          "description": "Returns an object containing a user's profile information. The path parameter (email) is required and must be an email that has been registered. This route returns additional information if the user is authenticated. To be authenticated, a valid JWT token must be sent in the header of the request. To test this using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in.",
          "parameters": [
            {
              "name": "email",
              "in": "path",
              "description": "Email address of a registered user",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "An object containing email, firstName, lastName, address (authenticated only) and dob (authenticated only) properties. An example object for an authenticated request is shown below. If no profile information has been entered for the user via the profile PUT request route, all returned values except email are expected to be null. Click on 'Schema' below to see a response example for an unauthenticated request.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/profileget200"
                  }
                }
              }
            },
            "401": {
                "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/profileget401"
                  }
                  }
                }
            },
            "404": {
                "description": "User not found",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/profileget404"
                  }
                  }
                }
            }
          },"security": [
            {
              "BearerAuth": [
                "write:pets",
                "read:pets"
              ]
            }
          ]
        }
      },
      "/user/{email}/profile ": {
        "put": {
          "tags": [
            "Profile"
          ],
          "description": "Updates a user's profile information. The path parameter (email) is required and must be an email that has been registered. A request body containing the user's profile information must be sent. This route returns an object containing the user's updated profile information. This route also requires the user to be authenticated - a valid JWT token must be sent in the header of the request. To test this route using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in.",
          "parameters": [
            {
              "name": "email",
              "in": "path",
              "description": "Email address of a registered user",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                
                "schema": {
                    
                    "$ref": "#/components/schemas/putrequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "An object containing email, firstName, lastName, address (authenticated only) and dob (authenticated only) properties. An example object for an authenticated request is shown below. If no profile information has been entered for the user via the profile PUT request route, all returned values except email are expected to be null. Click on 'Schema' below to see a response example for an unauthenticated request.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/putrequest"
                  }
                }
              }
            },
            "400": {
                "description": "Bad request. Click on 'Schema' below to see the possible error responses.",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/profileput400"
                  }
                  }
                }
            },
            "401": {
                "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/profileput401"
                  }
                  }
                }
            },
            "403": {
                "description": "Forbidden. Email address associated with JWT token is not the same as email provided in path parameter.",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/profileput403"
                  }
                  }
                }
            }
          },
          "security": [
            {
              "BearerAuth": [
                "write:pets",
                "read:pets"
              ]
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "register201":{
            
            "type":"object",
            "properties":{
                "message":{
                    "type":"string",
                    "example":"User created"
                }
            }
        },
...
        "profileput401":{
            "type":"object",
            "properties":{
                "error":{
                    "type":"boolean",
                    "example":true
                },
                "message":{
                    "type":"string",
                    "example":"Authorization header ('Bearer token') not found"
                }
            }
        },
...
        "star":{
            "type":"object",
            "properties":{
                "star":{
                    "type":"object",
                    "example":"user's name"
                } 
            }
        },
        "volcanoes":{
            "type":"object",
            "properties":{
                "id":{
                    "type":"integer",
                    "example":1
                },
                "name":{
                    "type":"string",
                    "example":"Abu"
                },
                "country":{
                    "type":"string",
                    "example":"Japan"
                },
                "region":{
                    "type":"string",
                    "example":"Japan, Taiwan, Marianas"
                },
                "subregion":{
                    "type":"string",
                    "example":"Honshu"
                }
                
                  
            }
        },
        "countries":{
            "type":"array",
            "items":{
                "type":"string",
                "example":[
                    "Algeria",
                    "Antarctica",
                    "Argentina",
                    "Armenia",
                    "Australia"
                ]
            }

        },
...
        "ErrorResponse":{
            "type":"object",
            "properties":{
                "error":{
                    "type":"boolean",
                    "example":true
                },
                "message":{
                    "type":"string",
                    "example":"Invalid query parameters. Query parameters are not permitted."
                }    
            }
        },
...
      },
      "securitySchemes": {
        
        "BearerAuth": {
          "type": "apiKey",
          "name": "Enter JWT Bearer token only",
          "in": "header"
        }
      }
    }
  }